Website | Gallery | Renaming basic scatter plot example #0
