+++ packages/shared/examples/_previews/sized-scatter-plot.png
