+++ packages/shared/examples/free-brush-scatters/data.ts
@@ -1,2 +1,2 @@
-export { data, palette } from '../basic-scatter-chart/data'
-export type { DataRecord } from '../basic-scatter-chart/data'
+export { data, palette } from '../sized-scatter-plot/data'
+export type { DataRecord } from '../sized-scatter-plot/data'
