+++ packages/shared/examples/free-brush-scatters/index.tsx
@@ -26,7 +26,7 @@ const example: Example = {
   },
   codeSvelte: require(`!!raw-loader!./${pathname}.svelte`).default,
   codeVue: require(`!!raw-loader!./${pathname}.vue`).default,
-  data: require('!!raw-loader!./../basic-scatter-chart/data.ts').default,
+  data: require('!!raw-loader!./../sized-scatter-plot/data.ts').default,
   preview: require(`../_previews/${pathname}.png`).default,
   previewDark: require(`../_previews/${pathname}-dark.png`).default,
   styles: require('!!raw-loader!./styles.css').default,
