+++ packages/shared/examples/sized-scatter-plot/data.ts
