+++ packages/shared/examples/sized-scatter-plot/sized-scatter-plot.svelte
